+++ pyrubrum/database/errors/not_found_error.py
@@ -16,9 +16,12 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with Pyrubrum. If not, see <http://www.gnu.org/licenses/>.
 
-from .delete_error import DeleteError  # noqa
-from .error import DatabaseError  # noqa
-from .expire_error import ExpireError  # noqa
-from .get_error import GetError  # noqa
-from .not_found_error import NotFoundError  # noqa
-from .set_error import SetError  # noqa
+from .error import DatabaseError
+
+
+class NotFoundError(DatabaseError):
+    """Exception which is raised whenever a key is not found inside the
+    database"""
+
+    def __init__(self, key: str):
+        super().__init__("%s was not found inside the database" % key)