+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 
 ### New features
 
-   - Import all the public functions
+   - Import all the public functions ([c0a7deb30fbe75d6b8c37f71dcaaa49ba8b3ab8f](https://github.com/hearot/pyrubrum/commit/c0a7deb30fbe75d6b8c37f71dcaaa49ba8b3ab8f))
 
 ### Other changes
 
@@ -25,6 +25,7 @@
 ### ‼️ Breaking changes
 
    - Create specific directories for module entities ([25593e6d40fa34dbc47528ff3fa6fdc30c0a41b9](https://github.com/hearot/pyrubrum/commit/25593e6d40fa34dbc47528ff3fa6fdc30c0a41b9))
+   - `BaseDatabase.get` always returns a string (resolve #2)
 
 ## v0.1a1.dev5 - 2020-06-12